Use ffmpeg's concat filter to concatenate videos
This commit adds an extra concat_videos_2 function that can concatenate
multiple input files, and not just two inputs. Currently, this is only
used when combining clips, but it can be used in other places, to
simplify the code.

The concatenation we were doing before seemed arbitrary and seemed to
cause issues like audio video lag, when adding background music or doing
thresholding of audio levels. This concatenation seems to fix some of
the problems.